move matrix upload to state flush
